api_gateway/requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.2
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
sqlparse==0.5.0
typing_extensions==4.12.2
